Add files via upload
The Dataset was downloaded from kaggle platform, it is a banking data-set which has information about customers of a bank. The bank has started a new campaign for it's term 
deposit plan and the main objective of the project is to predict whether a customer is going to subscribe to the term deposit or not by training the model on Prehistorical data provided by the bank.